Reset routes and templates to base state for demo
We want to start with a base demo. Note that there are other parts of the completed app in this branch. Those have no immediate bearing on what we are going to show in this demo and can be ignored. This demo is going to showcase how to use the router and how to create basic page views in ember with templates.